requirements-dev.txt
black==24.2.0
django-stubs==4.2.7
djlint==1.34.1
dnspython==2.6.1
dnstwist==20240116
ail-typo-squatting==2.7.4
